requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
elevenlabs==1.56.0
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.9.0
multidict==6.3.2
numpy==2.0.2
openai==1.70.0
pip==25.0.1
propcache==0.3.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
setuptools==58.1.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.13.1
typing-inspection==0.4.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==12.0
wsproto==1.2.0
yarl==1.18.3
